requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
streamlit
requests
Pillow
streamlit-drawable-canvas